Parameterize buffer for on-filetype
In preparation for the TODO added:

    TODO Call on-filetype for all open clojure buffers

Currently, if a plugin manager like Lazy loads this plugin only after a
Clojure file was loaded, then the buffer-local user commands will not be
created for that file.

Not sure when I'll get around to it, so workarounds are to either reload
any Clojure buffers after this plugin has been loaded, or to ensure that
this plugin isn't lazily loaded.